+++ previous-versions/sync-for-commerce-version-1/src/codatsynccommerce/sdkconfiguration.py
@@ -20,9 +20,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '1.1'
-    sdk_version: str = '0.3.0'
-    gen_version: str = '2.139.1'
-    user_agent: str = 'speakeasy-sdk/python 0.3.0 2.139.1 1.1 codat-sync-for-commerce-version-1'
+    sdk_version: str = '0.4.0'
+    gen_version: str = '2.143.2'
+    user_agent: str = 'speakeasy-sdk/python 0.4.0 2.143.2 1.1 codat-sync-for-commerce-version-1'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> tuple[str, dict[str, str]]: