+++ previous-versions/sync-for-commerce-version-1/RELEASES.md
@@ -29,3 +29,13 @@ Based on:
 - [python v0.3.0] previous-versions/sync-for-commerce-version-1
 ### Releases
 - [PyPI v0.3.0] https://pypi.org/project/codat-sync-for-commerce-version-1/0.3.0 - previous-versions/sync-for-commerce-version-1
+
+## 2023-10-02 00:16:15
+### Changes
+Based on:
+- OpenAPI Doc 1.1 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Commerce-v1.yaml
+- Speakeasy CLI 1.92.3 (2.143.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.4.0] previous-versions/sync-for-commerce-version-1
+### Releases
+- [PyPI v0.4.0] https://pypi.org/project/codat-sync-for-commerce-version-1/0.4.0 - previous-versions/sync-for-commerce-version-1